functions/requirements.txt
firebase-functions>=0.4.3
python-dotenv>=1.0.0
fastapi>=0.115.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
structlog>=24.4.0
google-cloud-translate>=3.15.0
google-cloud-language>=2.13.0
google-cloud-firestore>=2.16.0
